+++ shuttle/src/lib.rs
@@ -187,6 +187,7 @@ pub mod lazy_static;
 pub mod rand;
 pub mod sync;
 pub mod thread;
+pub mod time;
 
 pub mod current;
 pub mod scheduler;
